Add two actions to control sermon audio upload from the frontend
audioUploadWelcome: This is supposed to display a list of all sermons
without audio and allow to select one and attach an audio file to it.

audioUploadDone: This is supposed to add the audio file to the selected
sermon record and also tag (id3) the audio file with information from
the sermon record.